Team controller

+++ app/controllers/user_controller.py
@@ -22,6 +22,7 @@ def create_user():
 
     user = User.from_dict(data)
     user.password = generate_password_hash(user.password)
+    user.status = "active"
     result = mongo.db.users.insert_one(user.to_dict())
 
     return jsonify({
@@ -40,16 +41,13 @@ def login_user():
             "error" : "Faltan atributos"
         }), 400
         
-    
     user = mongo.db.users.find_one({"email": email})
-    
     
     if not user: 
         return jsonify({
             "error": "No se encontró a ningun usuario con ese correo"
         }), 400
         
-    
     if not check_password_hash(user["password"], password): 
        return jsonify({
             "error": "Las contraseñas son diferentes."
@@ -61,13 +59,48 @@ def login_user():
         "access_token": access_token
     })
     
+@user_bp.route("/update/status/<user_id>", methods=["PATCH"])
+def update_user_status(user_id): 
+    user = mongo.db.users.find_one({"_id": ObjectId(user_id)})
+    
+    if not user: 
+        return jsonify({
+            "error": "No se encontró un usuario"
+        })
+        
+    if user["status"] == "active": 
+        mongo.db.users.update_one({
+            "_id": ObjectId(user_id)},
+            {"$set": {"status": "inactive"}}
+        )
+        
+        return jsonify({"msg": "El estado se ha actualizado"}), 200
+    
+    if user["status"] == "inactive": 
+        mongo.db.users.update_one({
+            "_id": ObjectId(user_id)},
+            {"$set": {"status": "active"}}
+        )
+        
+        return jsonify({"msg": "El estado se ha actualizado"}), 200
+
+
+    
 @user_bp.route("/delete", methods=["DELETE"])
 @jwt_required()
 def delete_user(): 
     user_id = get_jwt_identity()
+    
+    user = mongo.db.users.find_one({"_id": ObjectId(user_id)})
+    
+    if user["status"] == "active": 
+        return jsonify({
+            "error": "No se puede eliminar a un usuario activo."
+        })
     
     mongo.db.users.delete_one({"_id": ObjectId(user_id)})
     
     return jsonify({
         "msg" : "Usuario eliminado."
     }), 200
+
